removes balances from transactionSubmission store
