+++ extension/src/helpers/metrics.ts
@@ -7,6 +7,7 @@ import { AMPLITUDE_KEY } from "constants/env";
 import { settingsDataSharingSelector } from "popup/ducks/settings";
 import { AccountType } from "@shared/api/types";
 import { captureException } from "@sentry/browser";
+import { METRIC_NAMES } from "popup/constants/metricsNames";
 
 type MetricHandler<AppState> = (state: AppState, action: AnyAction) => void;
 const handlersLookup: { [key: string]: MetricHandler<any>[] } = {};
@@ -160,4 +161,41 @@ export const emitMetric = async (name: string, body?: any) => {
     /* eslint-enable */
   });
   await uploadMetrics();
+};
+
+export const storeBalanceMetricData = (
+  publicKey: string,
+  accountFunded: boolean,
+) => {
+  const metricsData: MetricsData = JSON.parse(
+    localStorage.getItem(METRICS_DATA) || "{}",
+  );
+  const accountType = metricsData.accountType;
+
+  if (accountFunded && accountType === AccountType.HW) {
+    metricsData.hwFunded = true;
+  }
+  if (accountFunded && accountType === AccountType.IMPORTED) {
+    metricsData.importedFunded = true;
+  }
+  if (accountType === AccountType.FREIGHTER) {
+    // check if we found a previously unfunded freighter account for metrics
+    const unfundedFreighterAccounts =
+      metricsData.unfundedFreighterAccounts || [];
+    const idx = unfundedFreighterAccounts.indexOf(publicKey);
+
+    if (accountFunded) {
+      metricsData.freighterFunded = true;
+      if (idx !== -1) {
+        emitMetric(METRIC_NAMES.freighterAccountFunded, { publicKey });
+        unfundedFreighterAccounts.splice(idx, 1);
+      }
+    }
+    if (!accountFunded && idx === -1) {
+      unfundedFreighterAccounts.push(publicKey);
+    }
+    metricsData.unfundedFreighterAccounts = unfundedFreighterAccounts;
+  }
+
+  localStorage.setItem(METRICS_DATA, JSON.stringify(metricsData));
 };
